add CommentController update something

+++ modules/quhao.appserver/app/com/withiter/models/merchant/CommentEntityDef.java
@@ -9,10 +9,10 @@ public abstract class CommentEntityDef extends BaseModel {
 	public String uid;
 	public String mid;
 	public String averageCost = "0";
-	public int xingjiabi = 0;
-	public int kouwei = 0;
-	public int huanjing = 0;
-	public int fuwu = 0;
+	public float xingjiabi = 0;
+	public float kouwei = 0;
+	public float huanjing = 0;
+	public float fuwu = 0;
 
 	public String content = "";
